Refactor tests and add documentation
Pre-cursor to being able to support OBS -> OBS copying.
