+++ requirements-dev.txt
@@ -6,5 +6,5 @@ mock==1.3.0
 nose==1.3.7
 testfixtures==4.7.0
 coverage==4.4.1
-pytest==3.2.3
+pytest==3.3.1
 pytest-cov==2.5.1
