+++ setup.cfg
@@ -23,7 +23,7 @@ classifier =
 [coverage:run]
 branch = True
 source = stor
-omit = stor/tests/test_posix_path_compat.py,stor/tests/test_integration_swift.py,stor/tests/test_integration_s3.py,stor/tests/test_integration.py,stor/third_party/backoff.py
+omit = stor/tests/test_posix_path_compat.py,stor/tests/test_integration_swift.py,stor/tests/test_integration_s3.py,stor/tests/integration.py,stor/third_party/backoff.py
 
 [coverage:report]
 exclude_lines =
