+++ tox.ini
@@ -4,18 +4,22 @@ indexserver =
     default = https://pypi.python.org/simple/
 envlist = py27,py36
 
+[pytest]
+addopts = -v -rsx --ignore=stor/tests/integration.py
+
 [testenv]
 usedevelop = True
 setenv =
   XUNIT_FILE=nosetests-{envname}.xml
 commands =
     rm -f .coverage
-    py.test -v -rsx --cov=stor {posargs:stor}
+    py.test --cov=stor {posargs:stor}
 deps =
   -r{toxinidir}/requirements-dev.txt
 whitelist_externals = make
                       bash
                       nosetests
+                      rm
 passenv = SWIFT_TEST_USERNAME SWIFT_TEST_PASSWORD OS_TEMP_URL_KEY
           AWS_TEST_ACCESS_KEY_ID AWS_DEFAULT_REGION
           AWS_TEST_SECRET_ACCESS_KEY OS_AUTH_URL
